move PresetsInstall to its own header
